move dbConfig file in config directory
